requirements.txt
# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0

# MongoDB
pymongo>=4.6.0
dnspython>=2.4.0

# NLP & ML
torch>=2.0.0
transformers>=4.35.0
scikit-learn>=1.3.0
datasets>=2.15.0

# Text Matching
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Web App
streamlit>=1.29.0
